refactor(dto): separate response and request DTOs to different packages
