+++ src/main/java/kitchenpos/application/dto/request/CreateMenuGroupDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.request;
 
 public class CreateMenuGroupDto {
