+++ src/main/java/kitchenpos/application/dto/request/CreateTableGroupDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.request;
 
 import java.util.List;
 import org.springframework.util.CollectionUtils;
