+++ src/main/java/kitchenpos/application/dto/response/OrderDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.application.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
