+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
-import kitchenpos.application.dto.CreateMenuGroupDto;
-import kitchenpos.application.dto.MenuGroupDto;
+import kitchenpos.application.dto.request.CreateMenuGroupDto;
+import kitchenpos.application.dto.response.MenuGroupDto;
 import kitchenpos.domain.MenuGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
