+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -2,8 +2,8 @@ package kitchenpos.application;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.application.dto.CreateProductDto;
-import kitchenpos.application.dto.ProductDto;
+import kitchenpos.application.dto.request.CreateProductDto;
+import kitchenpos.application.dto.response.ProductDto;
 import kitchenpos.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
