+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -5,13 +5,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import kitchenpos.application.dto.CreateOrderDto;
-import kitchenpos.application.dto.CreateOrderLineItemDto;
-import kitchenpos.application.dto.CreateTableDto;
-import kitchenpos.application.dto.CreateTableGroupDto;
-import kitchenpos.application.dto.EmptyTableDto;
-import kitchenpos.application.dto.TableDto;
-import kitchenpos.application.dto.UpdateGuestNumberDto;
+import kitchenpos.application.dto.request.CreateOrderDto;
+import kitchenpos.application.dto.request.CreateOrderLineItemDto;
+import kitchenpos.application.dto.request.CreateTableDto;
+import kitchenpos.application.dto.request.CreateTableGroupDto;
+import kitchenpos.application.dto.request.EmptyTableDto;
+import kitchenpos.application.dto.response.TableDto;
+import kitchenpos.application.dto.request.UpdateGuestNumberDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
